feat: Add number formatting options in column editor with currency selection for enhanced data presentation

+++ server/src/rest/utils/column.ts
@@ -198,7 +198,7 @@ export const specificType = (
     case DataType.Integer:
       return tableBuilder.integer(name);
     case DataType.Number:
-      return tableBuilder.decimal(name);
+      return tableBuilder.double(name);
     case DataType.Checkbox:
       return tableBuilder.boolean(name);
     case DataType.Date:
